Created a useFileList hook
